#44 [MOD] 스트링 추출 및 크기 조정
