+++ presentation/src/main/java/com/sopt/presentation/calendar/CalendarRoute.kt
@@ -1,55 +1,30 @@
 package com.sopt.presentation.calendar
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.statusBarsPadding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.RectangleShape
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.dimensionResource
-import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.zIndex
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.sopt.core.designsystem.component.topappbar.NoostakTopAppBar
 import com.sopt.core.designsystem.theme.NoostakAndroidTheme
-import com.sopt.core.designsystem.theme.NoostakTheme
-import com.sopt.core.extension.noRippleClickable
 import com.sopt.domain.entity.CalendarGroupEntity
 import com.sopt.presentation.R
-import com.sopt.presentation.appointment.scrollToItem
 import com.sopt.presentation.calendar.component.CalendarFloatingActionDialog
-import com.sopt.presentation.calendar.component.CalendarGroupItem
+import com.sopt.presentation.calendar.component.CalendarGroup
 
 @Composable
 fun CalendarRoute(
@@ -100,11 +75,6 @@ fun CalendarScreen(
     groups: List<CalendarGroupEntity>,
     onAddBtnClick: () -> Unit = {}
 ) {
-    val coroutineScope = rememberCoroutineScope()
-    val listState = rememberLazyListState()
-    val density = LocalDensity.current
-    var selectedGroup by remember { mutableIntStateOf(0) }
-
     Scaffold(
         modifier = Modifier
             .fillMaxSize()
@@ -113,7 +83,7 @@ fun CalendarScreen(
             .navigationBarsPadding(),
         topBar = {
             NoostakTopAppBar(
-                title = "ScheTook",
+                title = stringResource(R.string.appbar_calendar),
                 isIconVisible = false,
                 isMainAppBar = true
             )
@@ -124,76 +94,10 @@ fun CalendarScreen(
                 .padding(innerPadding)
                 .padding(vertical = dimensionResource(id = R.dimen.vertical_padding))
         ) {
-            Box(
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Box(
-                    modifier = Modifier
-                        .padding(top = 5.dp, end = 16.dp)
-                        .size(46.dp)
-                        .background(
-                            color = NoostakTheme.colors.black,
-                            shape = CircleShape
-                        )
-                        .align(Alignment.TopEnd)
-                        .zIndex(2f)
-                        .noRippleClickable { onAddBtnClick() }
-                ) {
-                    Icon(
-                        modifier = Modifier
-                            .align(Alignment.Center)
-                            .size(20.dp),
-                        imageVector = ImageVector.vectorResource(R.drawable.ic_add),
-                        contentDescription = null,
-                        tint = NoostakTheme.colors.white
-                    )
-                }
-                Box(
-                    modifier = Modifier
-                        .zIndex(1f)
-                        .padding(end = 42.dp)
-                        .width(64.dp)
-                        .height(85.dp)
-                        .align(Alignment.CenterEnd)
-                        .background(
-                            brush = Brush.horizontalGradient(
-                                colors = listOf(
-                                    Color.Transparent,
-                                    NoostakTheme.colors.white
-                                )
-                            ),
-                            shape = RectangleShape
-                        )
-                )
-                LazyRow(
-                    state = listState,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(end = 42.dp),
-                    horizontalArrangement = Arrangement.spacedBy(10.dp)
-                ) {
-                    itemsIndexed(items = groups, key = { _, item -> item.id }) { index, group ->
-                        if (index == 0) {
-                            Spacer(
-                                modifier = Modifier.width(16.dp)
-                            )
-                        }
-                        CalendarGroupItem(
-                            data = group,
-                            isSelected = index == selectedGroup,
-                            onClick = {
-                                selectedGroup = index
-                                scrollToItem(listState, coroutineScope, density, index)
-                            }
-                        )
-                        if (index == groups.lastIndex) {
-                            Spacer(
-                                modifier = Modifier.width(41.dp)
-                            )
-                        }
-                    }
-                }
-            }
+            CalendarGroup(
+                groups = groups,
+                onAddBtnClick = onAddBtnClick
+            )
             Spacer(modifier = Modifier.height(8.dp))
         }
     }
@@ -217,7 +121,7 @@ fun CalendarScreenPreview() {
                 ),
                 CalendarGroupEntity(
                     id = 3,
-                    groupName = "유잔면",
+                    groupName = "유진면",
                     groupImage = "https://avatars.githubusercontent.com/u/91470334?v=4"
                 ),
                 CalendarGroupEntity(
